feat: layout View, Contoller ts 변환
